+++ nope/lib/views/user/form.php
@@ -14,7 +14,7 @@
     <div class="panel-body">
       <div class="form-group" ng-if="!user.id" ng-class="{'has-error':(!userForm.username.$valid && userForm.username.$touched)}">
         <label class="control-label">Username</label>
-        <input type="text" name="username" class="form-control" ng-model="user.username" required   ng-pattern='<?php echo \Nope\Utils::USERNAME_REGEX_PATTERN; ?>'' ng-trim="false" />
+        <input type="text" name="username" class="form-control" ng-model="user.username" required   ng-pattern='<?php echo \Nope\Utils::USERNAME_REGEX_PATTERN; ?>' ng-trim="false" placeholder="Username" />
         <div ng-messages="userForm.username.$error" ng-if="userForm.username.$touched" ng-cloak>
           <span class="help-block" ng-message="required">Username is required.</span>
           <span class="help-block" ng-message="pattern">Username must contain only lower case letters, numbers and it need to be between 3 and 20 chars long.</span>
@@ -25,7 +25,7 @@
           <div class="col col-md-6">
               <label class="control-label" ng-if="!user.id">Password</label>
               <label class="control-label" ng-if="user.id">New password</label>
-              <input type="password" name="password" class="form-control" ng-model="user.password" ng-required="!user.id" />
+              <input type="password" name="password" class="form-control" ng-model="user.password" ng-required="!user.id" placeholder="Password" />
               <div ng-messages="userForm.password.$error" ng-if="userForm.password.$touched" ng-cloak>
                 <span class="help-block" ng-message="required">Password is required.</span>
               </div>
@@ -33,7 +33,7 @@
           <div class="col col-md-6">
               <label class="control-label" ng-if="!user.id">Confirm password</label>
               <label class="control-label" ng-if="user.id">Confirm new password</label>
-              <input type="password" name="confirm" class="form-control" ng-model="user.confirm" ng-required="!user.id" nope-match="user.password" />
+              <input type="password" name="confirm" class="form-control" ng-model="user.confirm" ng-required="!user.id" nope-match="user.password" placeholder="Confirm password" />
               <div ng-messages="userForm.confirm.$error" ng-if="userForm.confirm.$touched" ng-cloak>
                 <span class="help-block" ng-message="required">Password confirmation is required.</span>
                 <span class="help-block" ng-message="match">Password and its confirmation must match.</span>
@@ -43,7 +43,7 @@
       </div>
       <div class="form-group" ng-class="{'has-error':(!userForm.email.$valid && userForm.email.$touched)}">
         <label class="control-label">Email</label>
-        <input type="text" ng-pattern='<?php echo \Nope\Utils::EMAIL_REGEX_PATTERN; ?>' name="email" class="form-control" ng-model="user.email" required />
+        <input type="text" ng-pattern='<?php echo \Nope\Utils::EMAIL_REGEX_PATTERN; ?>' name="email" class="form-control" ng-model="user.email" required placeholder="Email to reset password" />
         <div ng-messages="userForm.email.$error" ng-if="userForm.email.$touched" ng-cloak>
          <span class="help-block" ng-message="required">Email is required.</span>
          <span class="help-block" ng-message="pattern">Email must be valid.</span>
@@ -51,7 +51,7 @@
       </div>
       <div class="form-group">
         <label>Pretty name</label>
-        <input type="text" name="prettyname" class="form-control" ng-model="user.prettyName" />
+        <input type="text" name="prettyname" class="form-control" ng-model="user.prettyName"  placeholder="Pretty name to display" />
       </div>
       <div class="checkbox">
         <label>
@@ -61,13 +61,14 @@
       <div class="form-group">
         <label>Role</label>
         <select name="role" class="form-control" ng-model="user.role" required ng-if="currentUser.isAdmin() && !currentUser.itsMe(user)" nope-can="user.update" ng-disabled="!currentUser.isAdmin() || currentUser.itsMe(user)">
+          <option value="" disabled>Select role</option>
           <option ng-repeat="r in rolesList" value="{{r.key}}" ng-if="r.key!=='admin'">{{r.label}}</option>
         </select>
         <p class="form-control-static" ng-repeat="r in rolesList" ng-if="(!currentUser.isAdmin() || currentUser.isAdmin() && currentUser.itsMe(user)) && user.is(r.key)">{{r.label}}</p>
       </div>
       <div class="form-group">
         <label>Description</label>
-        <textarea name="description" class="form-control" ng-model="user.description"></textarea>
+        <textarea name="description" class="form-control" ng-model="user.description" placeholder="User brief description" rows="5"></textarea>
       </div>
       <div class="form-group">
         <label>Cover</label>