+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 ## v0.10.0 (2016-02-??)
 
+### Features
+- Added input fields placeholders (#27).
+
 ### Fixes
 - Fixed wrong `headline` setting creation during installation.
 - Fixed first gallery creation UI (#25).
 - Fixed user detail UI.
 - Fixed missing default fields (`type` and `provider`) on media upload.
+- Fixed reset password error without exception (#28).
 
 ## v0.9.0 (2016-02-03)